+++ packages/twenty-server/src/modules/connected-account/jobs/match-participant.job.ts
@@ -2,9 +2,7 @@ import { Injectable } from '@nestjs/common';
 
 import { MessageQueueJob } from 'src/engine/integrations/message-queue/interfaces/message-queue-job.interface';
 
-import { InjectObjectMetadataRepository } from 'src/engine/object-metadata-repository/object-metadata-repository.decorator';
-import { MessageParticipantRepository } from 'src/modules/messaging/repositories/message-participant.repository';
-import { MessageParticipantObjectMetadata } from 'src/modules/messaging/standard-objects/message-participant.object-metadata';
+import { MessageParticipantService } from 'src/modules/messaging/services/message-participant/message-participant.service';
 
 export type MatchParticipantJobData = {
   workspaceId: string;
@@ -18,36 +16,17 @@ export class MatchParticipantJob
   implements MessageQueueJob<MatchParticipantJobData>
 {
   constructor(
-    @InjectObjectMetadataRepository(MessageParticipantObjectMetadata)
-    private readonly messageParticipantRepository: MessageParticipantRepository,
+    private readonly messageParticipantService: MessageParticipantService,
   ) {}
 
   async handle(data: MatchParticipantJobData): Promise<void> {
-    const { workspaceId, personId, workspaceMemberId, email } = data;
+    const { workspaceId, email, personId, workspaceMemberId } = data;
 
-    const messageParticipantsToUpdate =
-      await this.messageParticipantRepository.getByHandles(
-        [email],
-        workspaceId,
-      );
-
-    const messageParticipantIdsToUpdate = messageParticipantsToUpdate.map(
-      (participant) => participant.id,
+    await this.messageParticipantService.matchMessageParticipants(
+      workspaceId,
+      email,
+      personId,
+      workspaceMemberId,
     );
-
-    if (personId) {
-      await this.messageParticipantRepository.updateParticipantsPersonId(
-        messageParticipantIdsToUpdate,
-        personId,
-        workspaceId,
-      );
-    }
-    if (workspaceMemberId) {
-      await this.messageParticipantRepository.updateParticipantsWorkspaceMemberId(
-        messageParticipantIdsToUpdate,
-        workspaceMemberId,
-        workspaceId,
-      );
-    }
   }
 }
