+++ packages/twenty-server/src/modules/messaging/services/message-participant/message-participant.module.ts
@@ -4,12 +4,16 @@ import { ObjectMetadataRepositoryModule } from 'src/engine/object-metadata-repos
 import { WorkspaceDataSourceModule } from 'src/engine/workspace-datasource/workspace-datasource.module';
 import { AddPersonIdAndWorkspaceMemberIdModule } from 'src/modules/connected-account/services/add-person-id-and-workspace-member-id/add-person-id-and-workspace-member-id.module';
 import { MessageParticipantService } from 'src/modules/messaging/services/message-participant/message-participant.service';
+import { MessageParticipantObjectMetadata } from 'src/modules/messaging/standard-objects/message-participant.object-metadata';
 import { PersonObjectMetadata } from 'src/modules/person/standard-objects/person.object-metadata';
 
 @Module({
   imports: [
     WorkspaceDataSourceModule,
-    ObjectMetadataRepositoryModule.forFeature([PersonObjectMetadata]),
+    ObjectMetadataRepositoryModule.forFeature([
+      PersonObjectMetadata,
+      MessageParticipantObjectMetadata,
+    ]),
     AddPersonIdAndWorkspaceMemberIdModule,
   ],
   providers: [MessageParticipantService],
